fix overflow

+++ src/main/scala/part1recap/ThreadModelLimitations.scala
@@ -87,7 +87,7 @@ object ThreadModelLimitations extends App {
   import scala.concurrent.ExecutionContext.Implicits.global
 
   val futures = (0 to 9)
-    .map(i => 100000 * i until 100000 * (i + 1)) // 0 - 99999, 100000 - 199999, 200000 - 299999 etc
+    .map(i => BigInt(100000 * i) until BigInt(100000 * (i + 1))) // 0 - 99999, 100000 - 199999, 200000 - 299999 etc
     .map(range => Future {
       if (range.contains(546735)) throw new RuntimeException("invalid number")
       range.sum
